+++ sig/herb/analysis/template_dependencies/local_scanner.rbs
@@ -1,7 +1,7 @@
-# Generated from lib/herb/action_view/template_dependencies/local_scanner.rb with RBS::Inline
+# Generated from lib/herb/analysis/template_dependencies/local_scanner.rb with RBS::Inline
 
 module Herb
-  module ActionView
+  module Analysis
     class TemplateDependencies
       class LocalScanner < ::Herb::Visitor
         attr_reader locals: untyped
